Move build-stamp and generatic files to obj/main/$(TARGET)
This allows target level parallel builds, since now there are no
shared files between targets.

Fixes #3261

+++ src/utils/build_stamp.rb
@@ -65,7 +65,9 @@ class Stamper
         output = File.join(@stamp_dir, "stamp")
         stdout, stderr = @compiler.run(input, output, ["-dM", "-E"])
         File.delete(input)
-        File.delete(output)
+        if File.file?(output)
+            File.delete(output)
+        end
         return Digest::SHA1.hexdigest(stdout)
     end
 end
